декомпозировал код, добавив хелперы для класса Kernel and Router
