+++ routes/web.php
@@ -8,5 +8,4 @@ use Artem\PhpFramework\Routing\Route;
 return [
     Route::get('/', [HomeController::class, 'index']),
     Route::get('/posts/{id}', [PostController::class, 'show']),
-
 ];
